// Client/TigerErpAngular-14/src/app/models/AccChartModel.ts
export class AccChartModel {
    id: number | null;
    lowerGroupId: number | null;
    groupAccountId: number | null;
    accountId: number | null;
    compId: number;
    branchId: number | null;
    accountName: string;
    compName: string;
    accountNo: string;
    accountNamed: string;
    aliasName: string;
    openningBalance: number | null;
    currencyId: number | null;
    accType: number | null;
    isSubledger: number | null;
    isBillbyBill: number | null;
    isInventory: number | null;
    isCostCenter: number | null;
    headGroupId: number | null;
    headGroupName: string;
    balanceType: number | null;
    isActive: number | null;
    depriciationRate: number | null;
    noteToHeadId: number | null;
    accountGroupId: number | null;
    accountGroupName: string;
    isSalesAccount: number | null;
    creditLimit: number | null;
    currencyRate: number | null;
    fCAcc: number | null;
    fCAmount: number | null;
    creditDays: number | null;
    isBothBill: number | null;
    printPorder: number | null;
    discountTypeId: number | null;
    commissionTypeId: number | null;
    isFixed: number | null;
    isIndependSubledger: number | null;
    subLeadeger: number | null;
    isEmployee: number | null;
    tIN: string;
    bIN: string;
    status: number | null;
}